DurationEffectController and PauseEffectController
