+++ packages/flame/lib/src/effects2/controllers/effect_controller.dart
@@ -81,3 +81,5 @@ abstract class EffectController {
     _forward = !_forward;
   }
 }
+
+
